refactoring flask application architecture
