+++ SpiderKeeper/app/proxy/__init__.py
@@ -1,0 +1,3 @@
+from SpiderKeeper.app.proxy.spiderctrl import SpiderAgent
+
+agent = SpiderAgent()
