+++ SpiderKeeper/app/templates/base.html
@@ -95,13 +95,13 @@
             </div>
             <ul class="sidebar-menu">
                 <li class="header">JOBS</li>
-                <li><a href="/project/{{ project.id }}/job/dashboard"><i
+                <li><a href="{{ url_for('dashboard.job_dashboard', project_id=project.id) }}"><i
                         class="fa fa-dashboard text-blue"></i>
                     <span>Dashboard</span></a></li>
-                <li><a href="/project/{{ project.id }}/job/periodic"><i class="fa fa-tasks text-green"></i>
+                <li><a href="{{ url_for('dashboard.job_periodic', project_id=project.id) }}"><i class="fa fa-tasks text-green"></i>
                     <span>Periodic Jobs</span></a></li>
                 <li class="header">SPIDERS</li>
-                <li><a href="/project/{{ project.id }}/spider/dashboard"><i class="fa fa-flask text-red"></i>
+                <li><a href="{{ url_for('dashboard.spider_dashboard', project_id=project.id) }}"><i class="fa fa-flask text-red"></i>
                     <span>Dashboard</span></a>
                 </li>
                 <li><a href="/project/{{ project.id }}/spider/deploy"><i class="fa fa-server text-orange"></i>
@@ -111,7 +111,7 @@
                 <li><a href="/project/{{ project.id }}/project/stats"><i class="fa  fa-area-chart text-gray"></i> <span>Running Stats</span></a>
                 </li>
                 <!--<li><a href="#"><i class="fa fa-circle-o text-red"></i> <span>Members</span></a></li>-->
-                <li><a href="{{ url_for('project_manage', project_id=project.id) }}"><i class="fa fa-gears text-red"></i> <span>Manage</span></a></li>
+                <li><a href="{{ url_for('dashboard.project_manage', project_id=project.id) }}"><i class="fa fa-gears text-red"></i> <span>Manage</span></a></li>
                 <li class="header">SERVER</li>
                 <li><a href="/project/{{ project.id }}/server/stats"><i class="fa fa-bolt text-red"></i> <span>Usage Stats</span></a>
                 </li>
