+++ SpiderKeeper/app/templates/job_dashboard.html
@@ -38,7 +38,7 @@
             {% if job.job_instance %}
             <tr>
                 <td>{{ job.job_execution_id }}</td>
-                <td><a href="{{ url_for('job_periodic', project_id=job.project_id) }}#{{ job.job_instance_id }}">{{ job.job_instance_id }}</a></td>
+                <td><a href="{{ url_for('dashboard.job_periodic', project_id=job.project_id) }}#{{ job.job_instance_id }}">{{ job.job_instance_id }}</a></td>
                 <td>{{ job.job_instance.spider_name }}</td>
                 <td class="txt-args" data-toggle="tooltip" data-placement="right"
                     title="{{ job.job_instance.spider_arguments }}">{{ job.job_instance.spider_arguments }}
@@ -95,7 +95,7 @@
             {% if job.job_instance %}
             <tr>
                 <td>{{ job.job_execution_id }}</td>
-                <td><a href="{{ url_for('job_periodic', project_id=job.project_id) }}#{{ job.job_instance_id }}">{{ job.job_instance_id }}</a></td>
+                <td><a href="{{ url_for('dashboard.job_periodic', project_id=job.project_id) }}#{{ job.job_instance_id }}">{{ job.job_instance_id }}</a></td>
                 <td>{{ job.job_instance.spider_name }}</td>
                 <td class="txt-args" data-toggle="tooltip" data-placement="right"
                     title="{{ job.job_instance.spider_arguments }}">{{ job.job_instance.spider_arguments }}
@@ -119,7 +119,7 @@
                 {% endif %}
                 <td>{{ timedelta(now,job.start_time) }}</td>
                 <td>{{ job.start_time }}</td>
-                <td><a href="{{ url_for('job_log', job_exec_id=job.job_execution_id) }}" target="_blank"
+                <td><a href="{{ url_for('dashboard.job_log', job_exec_id=job.job_execution_id) }}" target="_blank"
                        data-toggle="tooltip" data-placement="top" title="{{ job.service_job_execution_id }}">Log</a>
                 </td>
                 <td style="font-size: 10px;">{{ job.running_on }}</td>
@@ -163,7 +163,7 @@
             {% if job.job_instance %}
             <tr class="{% if job.has_errors %}danger{% elif job.has_warnings %}warning{% endif %}">
                 <td>{{ job.job_execution_id }}</td>
-                <td><a href="{{ url_for('job_periodic', project_id=job.project_id) }}#{{ job.job_instance_id }}">{{ job.job_instance_id }}</a></td>
+                <td><a href="{{ url_for('dashboard.job_periodic', project_id=job.project_id) }}#{{ job.job_instance_id }}">{{ job.job_instance_id }}</a></td>
                 <td>{{ job.job_instance.spider_name }}</td>
                 <td class="txt-args" data-toggle="tooltip" data-placement="right"
                     title="{{ job.job_instance.spider_arguments }}">{{ job.job_instance.spider_arguments }}
@@ -190,7 +190,7 @@
                 <td>{{ job.items_count }}</td>
                 <td>{{ job.warnings_count }}</td>
                 <td>{{ job.errors_count }}</td>
-                <td><a href="{{ url_for('job_log', job_exec_id=job.job_execution_id) }}" target="_blank"
+                <td><a href="{{ url_for('dashboard.job_log', job_exec_id=job.job_execution_id) }}" target="_blank"
                        data-toggle="tooltip" data-placement="top" title="{{ job.service_job_execution_id }}">Log</a>
                 </td>
                 {% if job.running_status == 2 %}
